uboot: Partition table: Inc userdata. Remove media
Existing media partition is not being used. Removing
this partition and increasing size of userdata partition
where are the media files are stored.

Fixed issue when partition size is greater than 4GB.
The sector size was being read into a u32 instead of u64.
While this will work for smaller sectors, it will not
work for sector sizes larger than 4GB.

Signed-off-by: Viswanath Puttagunta <vishp@ti.com>
Change-Id: Ifca02facd601785ff9848bcba75f5123a7914b57