uBoot: Modify fastboot download to for DMA to RAM transfer
The main issue was that for every 512 bytes transferred over DMA
the API was coming out of the DMA service just to memcpy these 512 bytes into a ram location.
Ancillary issue was that at the end of the transfer if the data
was less then the 512 we went into fifo transfer mode.
Not sure why since the DMA could handle small packets as long as it knew the size.

This is severely inefficient as memcpy is really expensive operation.

So when a download condition is detected go into DMA transfer mode and just direct the data
from DMA into the RAM, there by eliminating the memcpy altogether.
Then when fastboot indicated to take action on the downloaded data the RAM data was referenced directly.

Change-Id: Ie223f8d8caf15c4408647857850b82d0af227c99
Signed-off-by: Dan Murphy <dmurphy@ti.com>