uboot: Partition table: Add 16K crypto partition
This will create a new 16K partition that will be used for holding the
crypto footer used by Android's crypto implementation. This footer was
previously used in the efs partition, but efs changed to read-only to reduce
the risk of corrupting factory calibration data, thus the key storage must
move.

The Android crypto footer contains details on the type of encryption, and an
encrypted copy of the master key to decrypt the filesystem. Android uses
this footer without a filesystem, directly accessing the block device.

Change-Id: I80c0446b872ad8c9485e23c9848417f5bd49ee85
Signed-off-by: Bryan Buckley <bryan.buckley@ti.com>